fix(pr-comments): split components into separate files
