+++ examples/module1/lesson1/solver/components/calculatorInput.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export const InputField = ({ value, onChange }) => {
+  return <input
+    type='number'
+    className='rounded-md shadow-md p-4'
+    value={value}
+    onChange={onChange}
+  />;
+};
