+++ examples/module1/lesson1/solver/constant.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_VALUE = 0;
